+++ CHANGELOG.md
@@ -9,6 +9,13 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [2.0.1] 2024-01-23
+
+- Hid funktion selector by default.
+- Updated element key.
+- Handled search separately.
+- Added inherit values data type.
+
 ## [2.0.0] 2024-01-11
 
 - Updated element formatting to display subelement title.
@@ -82,7 +89,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Added OS2Forms organisation module
 - Added OS2Forms Organisation OpenID Connect module
 
-[Unreleased]: https://github.com/itk-dev/os2forms_organisation/compare/2.0.0...HEAD
+[Unreleased]: https://github.com/itk-dev/os2forms_organisation/compare/2.0.1...HEAD
+[2.0.1]: https://github.com/itk-dev/os2forms_organisation/compare/2.0.0...2.0.1
 [2.0.0]: https://github.com/itk-dev/os2forms_organisation/compare/1.3.3...2.0.0
 [1.3.3]: https://github.com/itk-dev/os2forms_organisation/compare/1.3.2...1.3.3
 [1.3.2]: https://github.com/itk-dev/os2forms_organisation/compare/1.3.1...1.3.2